// client/packages/editor-oss/src/controls/AiWorldController/AiWorldController.types.ts
export interface IAiTransformResponse {
    scale?: ITransformMessageData;
    rotation?: ITransformMessageData;
    position?: ITransformMessageData;
}